Rename FTPStream, support multiple line replies, add buffered read
Add basic Usage documentation.

Per Rust naming conventions, use CamelCase for types:
https://doc.rust-lang.org/style/style/naming/README.html

The documentation includes a basic example that is tested with `cargo
test`.

Update travis.yml to install pure-ftp server in Travis CI build

Use a BufReader around the TcpStream. Writes are not buffered, which I
think is actually best, because we want requests to be issued
immediately.