FIX - Re-add openssl for session re-use (#97)
Re-adds the openssl implementation with session re-use. Some servers require session re-use to be able to make the data connection. This is to protect from the possibility of an attacker hijacking the FTP data connection. native-tls doesn't support session re-use so openssl was added back to provide support for this.

Keeps native-tls feature gated. By keeping native-tls it allows Windows and OSX users to more easily use the library.

Adds From impl for FtpError to use ? in ftp.rs. This simplifies the code in ftp.rs making it cleaner and more maintainable.

Resolves #96